# apps/local-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexura Local Server API",
    "version": "0.0.1",
    "description": "The per-branch local server's real API surface -- LAN-only, offline-first (Auth/Distribution doc Part 2: this server has no public URL; it's reachable only on the hotel's own network). Every path below was extracted directly from the real Express route definitions in `src/routes/*.ts` (see `scripts/generate-openapi.mjs`), not hand-typed -- run `npm run openapi:generate` after changing routes to keep this in sync. Auth is a JWT in an httpOnly `access_token` cookie (12h, with an 8h offline-continuation grace period -- see `src/auth/tokens.ts`), not a bearer token in the traditional OpenAPI sense; represented here as a cookie API key for tooling compatibility.",
    "license": {
      "name": "Proprietary — internal use only"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local dev server"
    },
    {
      "url": "http://nexura.local:4000",
      "description": "Real branch deployment (Auth doc Part 2, Layer 3) -- LAN hostname, no public DNS/TLS. `localhost` above is genuinely correct for dev, not a placeholder; this second entry is the real production shape."
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  },
...
  "paths": {
...
    "/restaurant/menu": {
      "get": {
        "operationId": "getRestaurantMenu",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Menu, grouped by category",
        "description": "**Requires authentication** (any role).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        }
      }
    },
    "/restaurant/menu/categories": {
      "post": {
        "operationId": "postRestaurantMenuCategories",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Create a menu category",
        "description": "**Requires permission:** `restaurant:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/menu/items": {
      "post": {
        "operationId": "postRestaurantMenuItems",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Create a menu item",
        "description": "**Requires permission:** `restaurant:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "categoryId",
                  "name",
                  "price"
                ],
                "properties": {
                  "categoryId": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "exclusiveMinimum": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/menu/items/{id}/availability": {
      "post": {
        "operationId": "postRestaurantMenuItemsByIdAvailability",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "86 / re-enable a menu item",
        "description": "**Requires permission:** `restaurant:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "available"
                ],
                "properties": {
                  "available": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders": {
      "get": {
        "operationId": "getRestaurantOrders",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Order list",
        "description": "`?status=` filters (e.g. `open`, `sent_to_kitchen`, `served`, `closed`).\n\n**Requires permission:** `restaurant:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      },
      "post": {
        "operationId": "postRestaurantOrders",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Open an order (POS or Room Service)",
        "description": "Exactly one of `tableId` (POS) or `roomReservationId` (RT-06 room service) is required -- rejects with `400 INVALID_INPUT` if both or neither are set.\n\n**Requires permission:** `restaurant:operate`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tableId": {
                    "type": "string"
                  },
                  "roomReservationId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders/{id}": {
      "get": {
        "operationId": "getRestaurantOrdersById",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Order detail with items and running total",
        "description": "**Requires permission:** `restaurant:manage`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders/{id}/close": {
      "post": {
        "operationId": "postRestaurantOrdersByIdClose",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Close an order — post to room or take payment",
        "description": "`postToRoom: true` posts the total as a real folio charge on the linked reservation (refuses if the guest isn't checked in); otherwise `paymentMethod` is required.\n\n**Requires permission:** `restaurant:operate`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "postToRoom"
                ],
                "properties": {
                  "postToRoom": {
                    "type": "boolean"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "cash",
                      "card",
                      "transfer"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders/{id}/items": {
      "post": {
        "operationId": "postRestaurantOrdersByIdItems",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Add an item to an open order",
        "description": "**Requires permission:** `restaurant:operate`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "menuItemId"
                ],
                "properties": {
                  "menuItemId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders/{id}/items/{itemId}/status": {
      "post": {
        "operationId": "postRestaurantOrdersByIdItemsByItemIdStatus",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Update a single item's kitchen status (RT-02 KDS)",
        "description": "**Requires permission:** `restaurant:operate`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "ready",
                      "served"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/orders/{id}/send-to-kitchen": {
      "post": {
        "operationId": "postRestaurantOrdersByIdSendToKitchen",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Send order to the kitchen display",
        "description": "**Requires permission:** `restaurant:operate`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/room-charges": {
      "get": {
        "operationId": "getRestaurantRoomCharges",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Guest Room Charges (RT-07)",
        "description": "Restaurant-category folio charges only, read from the same `folio_charges` table Front Desk and Finance use.\n\n**Requires permission:** `folio:postcharge`",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          },
          "403": {
            "description": "Authenticated but lacks the required permission"
          }
        }
      }
    },
    "/restaurant/tables": {
      "get": {
        "operationId": "getRestaurantTables",
        "tags": [
          "Restaurant / POS"
        ],
        "summary": "Table Management board (RT-03)",
        "description": "**Requires authentication** (any role).",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Missing or invalid session"
          }
        }
      }
    },
...
  }
}